Moving types into ./src
