+++ src/api/mapping.ts
@@ -5,7 +5,7 @@ import util from 'util'
 import cp from 'child_process'
 import path from 'path'
 import { setData, getMappings } from '../lib/data'
-import { Mapping } from './types/general'
+import { Mapping } from '../types/general'
 const mappingRouter = express.Router()
 const exec = util.promisify(cp.exec)
