+++ src/providers/index.ts
@@ -1,4 +1,4 @@
-import { ServiceConfig } from '../api/types/general'
+import { ServiceConfig } from '../types/general'
 
 export default {
   dns_gd: require('./goDaddy')
