+++ src/types/general.ts
